css: Symbolic icons size must actually be specified relative to font size
Commit a277f8e0e1003b6c60d82f37f557e115a2781419 prevents icons from being
scaled when the global desktop resolution changes which is wrong since
symbolic icons are supposed to be treated like text.

Instead of specifying their size in pixels, go back to em but with a value
adjusted to the new font size that actually yields unscaled symbolic icons for
the default resolution.

https://bugzilla.gnome.org/show_bug.cgi?id=671657